apps/consumer-datalake/requirements.txt
kafka-python
python-dotenv
azure-storage-file-datalake